% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usas.R
\name{usas_accounts}
\alias{usas_accounts}
\title{Search federal accounts}
\usage{
usas_accounts(keywords = NULL, agency_id = NULL, max_results = 50)
}
\arguments{
\item{keywords}{Keyword filter}

\item{agency_id}{Agency identifier}

\item{max_results}{Max results (default 50)}
}
\value{
tibble of federal accounts with obligation/outlay totals
}
\description{
Search federal accounts
}
