% Generated by build script
\name{usas_agency}
\alias{usas_agency}
\title{Fetch agency overview details}
\usage{
usas_agency()
}
\arguments{
\item{toptier_code}{Agency toptier code (e.g. "012" for USDA, "097" for DOD).}
\item{fiscal_year}{Fiscal year (default: current)}
}
\value{tibble: one row with agency details}
\description{Fetch agency overview details}
