% Generated by build script
\name{usas_autocomplete_agency}
\alias{usas_autocomplete_agency}
\title{Autocomplete agency names}
\usage{
usas_autocomplete_agency()
}
\arguments{
\item{query}{Search text}
\item{limit}{Max results (default 10)}
}
\value{tibble: id, toptier_flag, subtier_agency_name, toptier_agency_name}
\description{Autocomplete agency names}
