% Generated by build script
\name{usas_autocomplete_naics}
\alias{usas_autocomplete_naics}
\title{Autocomplete NAICS codes}
\usage{
usas_autocomplete_naics()
}
\arguments{
\item{query}{Search text}
\item{limit}{Max results (default 10)}
}
\value{tibble: naics_code, naics_description}
\description{Autocomplete NAICS codes}
