% Generated by build script
\name{usas_autocomplete_recipient}
\alias{usas_autocomplete_recipient}
\title{Autocomplete recipient names}
\usage{
usas_autocomplete_recipient()
}
\arguments{
\item{query}{Search text}
\item{limit}{Max results (default 10)}
}
\value{tibble: recipient_name, uei, duns}
\description{Autocomplete recipient names}
