% Generated by build script
\name{usas_recipient}
\alias{usas_recipient}
\title{Fetch recipient details by hash ID}
\usage{
usas_recipient()
}
\arguments{
\item{recipient_id}{Recipient hash ID (from usas_recipients)}
\item{fiscal_year}{Optional fiscal year}
}
\value{tibble: one row with recipient overview}
\description{Fetch recipient details by hash ID}
