% Generated by build script
\name{usas_spending_time}
\alias{usas_spending_time}
\title{Spending over time}
\usage{
usas_spending_time()
}
\arguments{
\item{group}{Time grouping: "fiscal_year", "quarter", "month"}
\item{keywords}{Keyword filter}
\item{award_type}{Award type codes}
\item{agency}{Awarding agency name}
\item{fiscal_year}{Fiscal year filter}
}
\value{tibble: time_period, aggregated_amount}
\description{Spending over time}
