% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usas.R
\name{usas_states}
\alias{usas_states}
\title{List all recipient states with award amounts}
\usage{
usas_states()
}
\value{
tibble: fips, code, name, type, amount, count
}
\description{
List all recipient states with award amounts
}
