% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usas.R
\name{usas_recipient}
\alias{usas_recipient}
\title{Fetch recipient details by hash ID}
\usage{
usas_recipient(recipient_id, fiscal_year = NULL)
}
\arguments{
\item{recipient_id}{Recipient hash ID (from usas_recipients)}

\item{fiscal_year}{Optional fiscal year}
}
\value{
tibble: one row with recipient overview
}
\description{
Fetch recipient details by hash ID
}
