% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usas.R
\name{usas_spending_time}
\alias{usas_spending_time}
\title{Spending over time}
\usage{
usas_spending_time(
  group = "fiscal_year",
  keywords = NULL,
  award_type = c("A", "B", "C", "D"),
  agency = NULL,
  fiscal_year = NULL
)
}
\arguments{
\item{group}{Time grouping: "fiscal_year", "quarter", "month"}

\item{keywords}{Keyword filter}

\item{award_type}{Award type codes}

\item{agency}{Awarding agency name}

\item{fiscal_year}{Fiscal year filter}
}
\value{
tibble: time_period, aggregated_amount
}
\description{
Aggregated spending amounts grouped by time period.
}
